fix: refactor code for better organization
