+++ src/config.rs
@@ -3,7 +3,22 @@ use std::fs;
 use std::io;
 use std::path::{Path, PathBuf};
 
-use crate::cli::Config;
+/// Main config struct used throughout the app.
+#[derive(Debug, PartialEq, Clone)]
+pub struct Config {
+    pub directory: PathBuf,
+    pub output: PathBuf,
+    pub include_dirs: Option<Vec<String>>,
+    pub exclude_dirs: Option<Vec<String>>,
+    pub include_ext: Option<Vec<String>>,
+    pub exclude_ext: Option<Vec<String>>,
+    pub include_files: Option<Vec<String>>,
+    pub exclude_files: Option<Vec<String>>,
+    pub min_size: Option<u64>,
+    pub max_size: Option<u64>,
+    pub respect_gitignore: bool,
+    pub tree_only: bool,
+}
 
 /// Struct for deserializing YAML config file.
 #[derive(Debug, Serialize, Deserialize, Clone, Default)]
@@ -69,4 +84,98 @@ pub fn merge_config(file: FileConfig, cli: Config) -> Config {
         respect_gitignore: cli.respect_gitignore,
         tree_only: cli.tree_only,
     }
+}
+
+/// Create Config from clap ArgMatches
+pub fn config_from_matches(matches: clap::ArgMatches) -> std::io::Result<Config> {
+    let directory = matches
+        .get_one::<String>("directory")
+        .ok_or_else(|| std::io::Error::new(std::io::ErrorKind::InvalidInput, "Missing directory"))?
+        .into();
+    let output = matches
+        .get_one::<String>("output")
+        .ok_or_else(|| std::io::Error::new(std::io::ErrorKind::InvalidInput, "Missing output"))?
+        .into();
+
+    let include_dirs = matches.get_one::<String>("include_dirs").map(|dirs| {
+        dirs.split(',')
+            .map(|s| s.trim().to_lowercase())
+            .filter(|s| !s.is_empty())
+            .collect::<Vec<_>>()
+    });
+
+    let exclude_dirs = matches.get_one::<String>("exclude_dirs").map(|dirs| {
+        dirs.split(',')
+            .map(|s| s.trim().to_lowercase())
+            .filter(|s| !s.is_empty())
+            .collect::<Vec<_>>()
+    });
+
+    let include_ext = matches.get_one::<String>("include_ext").map(|ext| {
+        ext.split(',')
+            .map(|s| s.trim().to_lowercase())
+            .filter(|s| !s.is_empty())
+            .collect::<Vec<_>>()
+    });
+
+    let exclude_ext = matches.get_one::<String>("exclude_ext").map(|ext| {
+        ext.split(',')
+            .map(|s| s.trim().to_lowercase())
+            .filter(|s| !s.is_empty())
+            .collect::<Vec<_>>()
+    });
+
+    let include_files = matches.get_one::<String>("include_files").map(|files| {
+        files
+            .split(',')
+            .map(|s| s.trim().to_lowercase())
+            .filter(|s| !s.is_empty())
+            .collect::<Vec<_>>()
+    });
+
+    let exclude_files = matches.get_one::<String>("exclude_files").map(|files| {
+        files
+            .split(',')
+            .map(|s| s.trim().to_lowercase())
+            .filter(|s| !s.is_empty())
+            .collect::<Vec<_>>()
+    });
+
+    let min_size = matches
+        .get_one::<String>("min_size")
+        .map(|s| {
+            s.parse::<u64>().map_err(|_| {
+                std::io::Error::new(std::io::ErrorKind::InvalidInput, "Invalid min-size")
+            })
+        })
+        .transpose()?;
+    let max_size = matches
+        .get_one::<String>("max_size")
+        .map(|s| {
+            s.parse::<u64>().map_err(|_| {
+                std::io::Error::new(std::io::ErrorKind::InvalidInput, "Invalid max-size")
+            })
+        })
+        .transpose()?;
+    let respect_gitignore = matches
+        .get_one::<String>("respect_gitignore")
+        .map(|s| s == "true" || s == "1")
+        .unwrap_or(true);
+
+    let tree_only = matches.get_flag("tree_only");
+
+    Ok(Config {
+        directory,
+        output,
+        include_dirs,
+        exclude_dirs,
+        include_ext,
+        exclude_ext,
+        include_files,
+        exclude_files,
+        min_size,
+        max_size,
+        respect_gitignore,
+        tree_only,
+    })
 }
